requirements.txt
beautifulsoup4==4.14.2
chromadb==1.3.4
faiss_cpu==1.12.0
langchain==1.0.5
langchain_community==0.4.1
langchain_core==1.0.4
langchain_openai==1.0.2
openai==2.8.0
Requests==2.32.5
